chore: move ResultExtension into extensions folder
