+++ lib/core/extensions/future_result_extension.dart
