+++ lib/core/models/stowage_plan/container/pg_freight_containers.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:ext_rw/ext_rw.dart';
 import 'package:hmi_core/hmi_core.dart';
 import 'package:hmi_core/hmi_core_app_settings.dart';
-import 'package:sss_computing_client/core/future_result_extension.dart';
+import 'package:sss_computing_client/core/extensions/future_result_extension.dart';
 import 'package:sss_computing_client/core/models/stowage_plan/container/freight_container.dart';
 import 'package:sss_computing_client/core/models/stowage_plan/container/freight_containers.dart';
 import 'package:sss_computing_client/core/models/stowage_plan/container/json_freight_container.dart';
