+++ lib/core/models/stowage_plan/stowage_collection/pg_stowage_collection.dart
@@ -1,7 +1,7 @@
 import 'package:ext_rw/ext_rw.dart';
 import 'package:hmi_core/hmi_core.dart';
 import 'package:hmi_core/hmi_core_app_settings.dart';
-import 'package:sss_computing_client/core/future_result_extension.dart';
+import 'package:sss_computing_client/core/extensions/future_result_extension.dart';
 import 'package:sss_computing_client/core/models/stowage_plan/container/freight_container.dart';
 import 'package:sss_computing_client/core/models/stowage_plan/slot/slot.dart';
 import 'package:sss_computing_client/core/models/stowage_plan/slot/standard_slot.dart';
