+++ lib/core/models/voyage/pg_waypoints.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:ext_rw/ext_rw.dart';
 import 'package:hmi_core/hmi_core.dart';
 import 'package:hmi_core/hmi_core_app_settings.dart';
-import 'package:sss_computing_client/core/future_result_extension.dart';
+import 'package:sss_computing_client/core/extensions/future_result_extension.dart';
 import 'package:sss_computing_client/core/models/voyage/json_waypoint.dart';
 import 'package:sss_computing_client/core/models/voyage/waypoint.dart';
 import 'package:sss_computing_client/core/models/voyage/waypoints.dart';
