+++ lib/core/models/weight/pg_displacement_weights.dart
@@ -1,7 +1,7 @@
 import 'package:ext_rw/ext_rw.dart';
 import 'package:hmi_core/hmi_core.dart';
 import 'package:hmi_core/hmi_core_app_settings.dart';
-import 'package:sss_computing_client/core/future_result_extension.dart';
+import 'package:sss_computing_client/core/extensions/future_result_extension.dart';
 import 'package:sss_computing_client/core/models/weight/displacement_weight.dart';
 import 'package:sss_computing_client/core/models/weight/displacement_weights.dart';
 import 'package:sss_computing_client/core/models/weight/json_displacement_weight.dart';
