+++ lib/presentation/loading/widgets/cargo_column/cargo_weight_column.dart
@@ -4,7 +4,9 @@ import 'package:hmi_widgets/hmi_widgets.dart';
 import 'package:sss_computing_client/core/models/cargo/cargo.dart';
 import 'package:sss_computing_client/core/models/cargo/json_cargo.dart';
 import 'package:sss_computing_client/core/models/field/field_type.dart';
+import 'package:sss_computing_client/core/models/number_math_relation/greater_than_or_equal_to.dart';
 import 'package:sss_computing_client/core/models/record/value_record.dart';
+import 'package:sss_computing_client/core/validation/number_math_relation_validation_case.dart';
 import 'package:sss_computing_client/core/validation/real_validation_case.dart';
 import 'package:sss_computing_client/core/validation/required_validation_case.dart';
 import 'package:sss_computing_client/core/widgets/table/table_column.dart';
@@ -64,10 +66,16 @@ class CargoWeightColumn implements TableColumn<Cargo, double?> {
   bool get isResizable => true;
   //
   @override
-  Validator? get validator => const Validator(
+  Validator? get validator => Validator(
         cases: [
-          RequiredValidationCase(),
-          RealValidationCase(),
+          const RequiredValidationCase(),
+          const RealValidationCase(),
+          NumberMathRelationValidationCase(
+            relation: const GreaterThanOrEqualTo(),
+            secondValue: 0.0,
+            toValue: (text) => parseToValue(text) ?? 0.0,
+            customMessage: const Localized('weight cannot be negative').v,
+          ),
         ],
       );
   //
